[feat] (model)添加了hava方法通过邀请码id判断邀请码存不存在

+++ src/main/java/com/hniecs/mainserver/service/admin/InvitationCodeService.java
@@ -163,6 +163,9 @@ public class InvitationCodeService {
      */
     public String deleteById(Long id) {
         // TODO 判断该id邀请码是否存在
+        if(!invitationCodeModel.have(id)){
+            return "邀请码不存在";
+        }
         return invitationCodeModel.deleteById(id);
     }
 
@@ -172,6 +175,13 @@ public class InvitationCodeService {
      */
     public String updateInvitationCode(InvitationCodeEntity invitationCode) {
         // TODO 判断该邀请码实体是否存在
+        Long id=invitationCode.getId();
+        if (id==null){
+            return "邀请码不存在";
+        }
+        if(!invitationCodeModel.have(id)){
+            return "邀请码实体不存在";
+        }
         return invitationCodeModel.updateInvitationCode(invitationCode);
     }
 }
